+++ src/main/java/br/com/farmacia/servlet/CadastrarProduto.java
@@ -25,7 +25,7 @@ public class CadastrarProduto extends HttpServlet {
 		// se tirar o id_Prod está dando erro na contrutor
 		int id_Produto = Integer.parseInt(request.getParameter("id_Produto"));
 		String nome_Prod = request.getParameter("nome_Prod");
-		double preco_Prod = Double.parseDouble(request.getParameter("preco_Prod"));
+		float preco_Prod = Float.parseFloat(request.getParameter("preco_Prod"));
 		String descricao_Prod = request.getParameter("descricao_Prod");
 		int qtd_Prod = Integer.parseInt(request.getParameter("qtd_Prod"));
